Daily package downloads from the RStudio CRAN mirror

Daily package downloads from the RStudio CRAN mirror

cran_downloads(
  packages = NULL,
  when = c("last-day", "last-week", "last-month"),
  from = "last-day",
  to = "last-day",
  batch_size = 800,
  mc.cores = 1
)

Arguments

packages

A character vector, the packages to query, or NULL for a sum of downloads for all packages. Alternatively, it can also be "R", to query downloads of R itself. "R" cannot be mixed with packages.

when

last-day, last-week or last-month (see details). If this is given, then from and to are ignored.

from

Start date, in yyyy-mm-dd format, or last-day. It is ignored if when is given.

to

End date, in yyyy-mm-dd format, or last-day. It is ignored if when is given.

batch_size

Maximum batch size per query. A batch size <=800 is reccomended, as going above that can cause errors.

mc.cores

Number of cores to parallelise batched queries across.

Value

For packages a data frame with columns:

package

The package. This column is missing if all packages were queried.

date

Day of the downloads, it is a Date object.

count

Download count.

For downloads of R, there are also columns for the operating system (os) and the R version (version).

Details

last-day is the last day for which data is available, last-week is from 6 days prior to that last day with data, last-month is from 29 days prior to that last day with data.

0 counts can be due to the non-availability of data on the RStudio server for that day.

See also

Examples

## Default is last day for which data is available.
cran_downloads()
#>         date   count
#> 1 2022-12-18 2996182

## All downloads for 'dplyr' in the last day for which data is available.
cran_downloads(packages = "dplyr")
#>         date count package
#> 1 2022-12-18 31816   dplyr

## Daily downloads for 'igraph' last week
## (6 days prior to the last day for which data is available)
cran_downloads(packages = "igraph", when = "last-week")
#>         date count package
#> 1 2022-12-12  9778  igraph
#> 2 2022-12-13 10288  igraph
#> 3 2022-12-14  9310  igraph
#> 4 2022-12-15  8850  igraph
#> 5 2022-12-16  7521  igraph
#> 6 2022-12-17  4228  igraph
#> 7 2022-12-18  4108  igraph

## Downloads in the specified time interval
cran_downloads(from = "2014-06-30", to = "2014-08-08")
#>          date  count
#> 1  2014-06-30 211630
#> 2  2014-07-01 167871
#> 3  2014-07-02 145978
#> 4  2014-07-03 156735
#> 5  2014-07-04 121388
#> 6  2014-07-05  85680
#> 7  2014-07-06  88702
#> 8  2014-07-07 173305
#> 9  2014-07-08 225455
#> 10 2014-07-09 204074
#> 11 2014-07-10 196067
#> 12 2014-07-11 186799
#> 13 2014-07-12 128945
#> 14 2014-07-13 100936
#> 15 2014-07-14 186627
#> 16 2014-07-15 194915
#> 17 2014-07-16 181111
#> 18 2014-07-17 191137
#> 19 2014-07-18 173551
#> 20 2014-07-19 109991
#> 21 2014-07-20 103021
#> 22 2014-07-21 206653
#> 23 2014-07-22 177675
#> 24 2014-07-23 186140
#> 25 2014-07-24 150904
#> 26 2014-07-25 160480
#> 27 2014-07-26 100130
#> 28 2014-07-27  82237
#> 29 2014-07-28 142225
#> 30 2014-07-29 151380
#> 31 2014-07-30 176169
#> 32 2014-07-31 167655
#> 33 2014-08-01 238507
#> 34 2014-08-02  88788
#> 35 2014-08-03  89176
#> 36 2014-08-04 240695
#> 37 2014-08-05 311740
#> 38 2014-08-06 237572
#> 39 2014-08-07 229844
#> 40 2014-08-08 181458

## Multiple packages
cran_downloads(packages = c("ggplot2", "plyr", "dplyr"))
#>         date count package
#> 1 2022-12-18 61663 ggplot2
#> 2 2022-12-18  8654    plyr
#> 3 2022-12-18 31816   dplyr

## R downloads
cran_downloads("R") 
#>          date      version  os count
#> 1  2022-12-18        3.4.0 src     1
#> 2  2022-12-18        4.0.2 src     1
#> 3  2022-12-18        3.2.4 osx     2
#> 4  2022-12-18     3.6.3.nn osx    33
#> 5  2022-12-18        3.5.3 osx     4
#> 6  2022-12-18        3.3.3 osx    32
#> 7  2022-12-18        4.2.0 win    27
#> 8  2022-12-18        4.0.4 win     3
#> 9  2022-12-18        3.3.2 osx     1
#> 10 2022-12-18        4.1.3 src     6
#> 11 2022-12-18        4.1.2 osx  1179
#> 12 2022-12-18       latest src     1
#> 13 2022-12-18        4.1.0 win     9
#> 14 2022-12-18        3.6.0 win     3
#> 15 2022-12-18        3.4.1 src     1
#> 16 2022-12-18        4.2.1 win    16
#> 17 2022-12-18        3.6.3 osx    50
#> 18 2022-12-18       2.15.1 osx     1
#> 19 2022-12-18        4.0.1 win     1
#> 20 2022-12-18        3.6.2 osx     1
#> 21 2022-12-18        4.0.5 win     9
#> 22 2022-12-18        3.4.4 win     2
#> 23 2022-12-18        4.1.1 win     5
#> 24 2022-12-18        4.1.0 osx     4
#> 25 2022-12-18        3.2.1 osx     4
#> 26 2022-12-18        4.2.1 osx    19
#> 27 2022-12-18        2.3.0 src     1
#> 28 2022-12-18       2.12.1 osx     1
#> 29 2022-12-18       2.12.2 src     1
#> 30 2022-12-18        3.6.3 win   122
#> 31 2022-12-18      release osx     9
#> 32 2022-12-18        3.6.2 win     1
#> 33 2022-12-18        4.0.5 osx     6
#> 34 2022-12-18        3.4.4 osx     4
#> 35 2022-12-18  patched.exe win     1
#> 36 2022-12-18        4.1.1 osx    64
#> 37 2022-12-18        3.4.2 src     1
#> 38 2022-12-18        3.5.3 win     2
#> 39 2022-12-18        2.7.0 src     1
#> 40 2022-12-18        4.2.2 src   330
#> 41 2022-12-18        4.2.0 osx     3
#> 42 2022-12-18        4.0.4 osx     2
#> 43 2022-12-18        4.0.0 win     3
#> 44 2022-12-18        4.1.2 win     4
#> 45 2022-12-18        3.3.0 src     1
#> 46 2022-12-18        3.4.3 win     2
#> 47 2022-12-18        4.0.2 win     3
#> 48 2022-12-18       2.14.1 osx     1
#> 49 2022-12-18        4.2.2 osx  4898
#> 50 2022-12-18        4.2.0 src    15
#> 51 2022-12-18        4.1.3 win   831
#> 52 2022-12-18        3.6.1 win     4
#> 53 2022-12-18        3.2.5 src     1
#> 54 2022-12-18       2.13.1 osx     3
#> 55 2022-12-18        3.4.1 win     1
#> 56 2022-12-18        4.2.1 src     8
#> 57 2022-12-18        3.3.1 osx     1
#> 58 2022-12-18       2.10.1 osx     1
#> 59 2022-12-18        4.0.1 src     1
#> 60 2022-12-18 4.2.2patched win    17
#> 61 2022-12-18        4.1.1 src     5
#> 62 2022-12-18       2.11.0 osx     1
#> 63 2022-12-18       latest osx     1
#> 64 2022-12-18        3.6.3 src     4
#> 65 2022-12-18        devel win   798
#> 66 2022-12-18        4.0.3 win    35
#> 67 2022-12-18        3.5.2 src     4
#> 68 2022-12-18        3.4.2 win     1
#> 69 2022-12-18        4.0.2 osx     6
#> 70 2022-12-18        3.2.4 src     1
#> 71 2022-12-18        3.5.3 src     3
#> 72 2022-12-18       2.11.1 osx     1
#> 73 2022-12-18        4.2.2 win 47464
#> 74 2022-12-18        4.0.0 src     2
#> 75 2022-12-18        4.1.3 osx   715
#> 76 2022-12-18        4.1.2 src     4
#> 77 2022-12-18        3.6.1 osx     1
#> 78 2022-12-18        3.1.3 src     1