Top downloaded packages from the RStudio CRAN mirror

Top downloaded packages from the RStudio CRAN mirror

cran_top_downloads(when = c("last-day", "last-week", "last-month"), count = 10)

Arguments

when

last-day, last-week or last-month (see details).

count

Number of packages to list. Note that the DB server lists only at most 100 packages. This number might change in the future.

Value

A data frame with columns: rank, package,

count, from, to.

Details

last-day is the last day for which data is available, last-week is from 6 days prior to that last day with data, last-month is from 29 days prior to that last day with data.

0 counts can be due to the non-availability of data on the RStudio server for that day.

See also

Examples

if (FALSE) {
## Default is last day for which data is available.
cran_top_downloads()

## Last week (6 days prior to the last day for which data is available) 
## instead
cran_top_downloads(when = "last-week")
}