Daily package downloads from the RStudio CRAN mirror

Daily package downloads from the RStudio CRAN mirror

cran_downloads(packages = NULL, when = c("last-day", "last-week",
  "last-month"), from = "last-day", to = "last-day")

Arguments

packages
A character vector, the packages to query, or NULL for a sum of downloads for all packages. Alternatively, it can also be "R", to query downloads of R itself. "R" cannot be mixed with packages.

when
last-day, last-week or last-month. If this is given, then from and to are ignored.

from
Start date, in yyyy-mm-dd format, or last-day. It is ignored if when is given.

to
End date, in yyyy-mm-dd format, or last-day. It is ignored if when is given.

Value

For packages a data frame with columns:

package

The package. This column is missing if all packages were queried.

date

Day of the downloads, it is a Date object.

count

Download count.

For downloads of R, there are also columns for the operating system (os) and the R version (version).
See also

Examples

## All downloads yesterday
cran_downloads()

#>         date   count
#> 1 2019-04-09 3509140

## All downloads for 'dplyr' yesterday
cran_downloads(packages = "dplyr")

#>         date count package
#> 1 2019-04-09 34267   dplyr

## Daily downloads for 'igraph' last week
cran_downloads(packages = "igraph", when = "last-week")

#>         date count package
#> 1 2019-04-03 15981  igraph
#> 2 2019-04-04 15462  igraph
#> 3 2019-04-05 15016  igraph
#> 4 2019-04-06 11925  igraph
#> 5 2019-04-07 11868  igraph
#> 6 2019-04-08 15802  igraph
#> 7 2019-04-09 15935  igraph

## Downloads in the specified time interval
cran_downloads(from = "2014-06-30", to = "2014-08-08")

#>          date  count
#> 1  2014-06-30 211630
#> 2  2014-07-01 167871
#> 3  2014-07-02 145978
#> 4  2014-07-03 156735
#> 5  2014-07-04 121388
#> 6  2014-07-05  85680
#> 7  2014-07-06  88702
#> 8  2014-07-07 173305
#> 9  2014-07-08 225455
#> 10 2014-07-09 204074
#> 11 2014-07-10 196067
#> 12 2014-07-11 186799
#> 13 2014-07-12 128945
#> 14 2014-07-13 100936
#> 15 2014-07-14 186627
#> 16 2014-07-15 194915
#> 17 2014-07-16 181111
#> 18 2014-07-17 191137
#> 19 2014-07-18 173551
#> 20 2014-07-19 109991
#> 21 2014-07-20 103021
#> 22 2014-07-21 206653
#> 23 2014-07-22 177675
#> 24 2014-07-23 186140
#> 25 2014-07-24 150904
#> 26 2014-07-25 160480
#> 27 2014-07-26 100130
#> 28 2014-07-27  82237
#> 29 2014-07-28 142225
#> 30 2014-07-29 151380
#> 31 2014-07-30 176169
#> 32 2014-07-31 167655
#> 33 2014-08-01 238507
#> 34 2014-08-02  88788
#> 35 2014-08-03  89176
#> 36 2014-08-04 240695
#> 37 2014-08-05 311740
#> 38 2014-08-06 237572
#> 39 2014-08-07 229844
#> 40 2014-08-08 181458

## Multiple packages
cran_downloads(packages = c("ggplot2", "plyr", "dplyr"))

#>         date count package
#> 1 2019-04-09 38956 ggplot2
#> 2 2019-04-09 20269    plyr
#> 3 2019-04-09 34267   dplyr

## R downloads
cran_downloads("R")

#>          date      version  os count
#> 1  2019-04-09       latest  NA    10
#> 2  2019-04-09       2.10.0 osx     3
#> 3  2019-04-09       2.10.1 osx     4
#> 4  2019-04-09       2.11.0 osx     4
#> 5  2019-04-09       2.11.1 osx     3
#> 6  2019-04-09       2.12.0 osx     4
#> 7  2019-04-09       2.12.1 osx     4
#> 8  2019-04-09       2.12.2 osx     3
#> 9  2019-04-09       2.13.0 osx     3
#> 10 2019-04-09       2.13.1 osx     2
#> 11 2019-04-09       2.13.2 osx     1
#> 12 2019-04-09       2.14.0 osx     1
#> 13 2019-04-09       2.14.1 osx     2
#> 14 2019-04-09       2.14.2 osx     2
#> 15 2019-04-09       2.15.0 osx     2
#> 16 2019-04-09       2.15.1 osx     4
#> 17 2019-04-09       2.15.2 osx     3
#> 18 2019-04-09       2.15.3 osx     3
#> 19 2019-04-09        3.0.0 osx     4
#> 20 2019-04-09        3.0.1 osx     3
#> 21 2019-04-09        3.0.2 osx     2
#> 22 2019-04-09        3.0.3 osx     2
#> 23 2019-04-09        3.1.0 osx     6
#> 24 2019-04-09        3.1.1 osx     2
#> 25 2019-04-09        3.1.2 osx     5
#> 26 2019-04-09        3.1.3 osx     4
#> 27 2019-04-09        3.2.0 osx     2
#> 28 2019-04-09        3.2.1 osx    12
#> 29 2019-04-09        3.2.2 osx     2
#> 30 2019-04-09        3.2.3 osx     1
#> 31 2019-04-09        3.2.4 osx    18
#> 32 2019-04-09        3.3.0 osx     1
#> 33 2019-04-09        3.3.1 osx     2
#> 34 2019-04-09        3.3.2 osx     2
#> 35 2019-04-09        3.3.3 osx    58
#> 36 2019-04-09        3.4.0 osx     7
#> 37 2019-04-09        3.4.1 osx     5
#> 38 2019-04-09        3.4.2 osx     2
#> 39 2019-04-09        3.4.3 osx     3
#> 40 2019-04-09        3.4.4 osx    77
#> 41 2019-04-09        3.5.0 osx     3
#> 42 2019-04-09        3.5.1 osx     5
#> 43 2019-04-09        3.5.2 osx     8
#> 44 2019-04-09        3.5.3 osx   732
#> 45 2019-04-09       latest osx   475
#> 46 2019-04-09       2.15.0 src     1
#> 47 2019-04-09       2.15.2 src     2
#> 48 2019-04-09        3.0.0 src     1
#> 49 2019-04-09        3.0.3 src     3
#> 50 2019-04-09        3.2.2 src    13
#> 51 2019-04-09        3.2.3 src     1
#> 52 2019-04-09        3.2.5 src     1
#> 53 2019-04-09        3.3.0 src     1
#> 54 2019-04-09        3.3.2 src     1
#> 55 2019-04-09        3.3.3 src    58
#> 56 2019-04-09        3.4.0 src     2
#> 57 2019-04-09        3.4.1 src    11
#> 58 2019-04-09        3.4.2 src     4
#> 59 2019-04-09        3.4.3 src     4
#> 60 2019-04-09        3.4.4 src     3
#> 61 2019-04-09        3.5.0 src     3
#> 62 2019-04-09        3.5.1 src    30
#> 63 2019-04-09        3.5.2 src    10
#> 64 2019-04-09        3.5.3 src   143
#> 65 2019-04-09       latest src     3
#> 66 2019-04-09       2.13.1 win     1
#> 67 2019-04-09       2.13.2 win     1
#> 68 2019-04-09       2.15.1 win     1
#> 69 2019-04-09        3.0.0 win     1
#> 70 2019-04-09        3.0.1 win     1
#> 71 2019-04-09        3.0.2 win     1
#> 72 2019-04-09        3.0.3 win     2
#> 73 2019-04-09        3.1.0 win     3
#> 74 2019-04-09        3.1.1 win     1
#> 75 2019-04-09        3.1.3 win     1
#> 76 2019-04-09        3.2.0 win     1
#> 77 2019-04-09        3.2.1 win     6
#> 78 2019-04-09        3.2.5 win     9
#> 79 2019-04-09        3.3.0 win     2
#> 80 2019-04-09        3.3.1 win     1
#> 81 2019-04-09        3.3.2 win     1
#> 82 2019-04-09        3.3.3 win    41
#> 83 2019-04-09        3.4.2 win     2
#> 84 2019-04-09        3.4.3 win    19
#> 85 2019-04-09        3.4.4 win    46
#> 86 2019-04-09        3.5.0 win    27
#> 87 2019-04-09        3.5.1 win     4
#> 88 2019-04-09        3.5.2 win     1
#> 89 2019-04-09        3.5.3 win  4168
#> 90 2019-04-09 3.5.3patched win   464
#> 91 2019-04-09   3.6.0alpha win     1
#> 92 2019-04-09        devel win    83