Top downloaded packages from the RStudio CRAN mirror

Top downloaded packages from the RStudio CRAN mirror

cran_top_downloads(when = c("last-day", "last-week", "last-month"),
  count = 10)

Arguments

when
last-day, last-week or last-month (see details).

count
Number of packages to list. Note that the DB server lists only at most 100 packages. This number might change in the future.

Value

A data frame with columns: rank, package, count, from, to.

Details

last-day is the last day for which data is available, last-week is from 6 days prior to that last day with data, last-month is from 29 days prior to that last day with data.

0 counts can be due to the non-availability of data on the RStudio server for that day.

See also

Examples

# \dontrun{
## Default is last day for which data is available.
cran_top_downloads()

#>    rank         package  count       from         to
#> 1     1        magrittr 120414 2019-12-01 2019-12-01
#> 2     2          aws.s3 111274 2019-12-01 2019-12-01
#> 3     3 aws.ec2metadata 110367 2019-12-01 2019-12-01
#> 4     4       rsconnect 106929 2019-12-01 2019-12-01
#> 5     5           rlang  29510 2019-12-01 2019-12-01
#> 6     6           dplyr  26430 2019-12-01 2019-12-01
#> 7     7         ggplot2  22036 2019-12-01 2019-12-01
#> 8     8            Rcpp  21629 2019-12-01 2019-12-01
#> 9     9        ellipsis  20457 2019-12-01 2019-12-01
#> 10   10           tidyr  20035 2019-12-01 2019-12-01

## Last week (6 days prior to the last day for which data is available) 
## instead
cran_top_downloads(when = "last-week")

#>    rank         package  count       from         to
#> 1     1        magrittr 891917 2019-11-25 2019-12-01
#> 2     2          aws.s3 768456 2019-11-25 2019-12-01
#> 3     3 aws.ec2metadata 761264 2019-11-25 2019-12-01
#> 4     4       rsconnect 740251 2019-11-25 2019-12-01
#> 5     5           rlang 356529 2019-11-25 2019-12-01
#> 6     6           dplyr 259152 2019-11-25 2019-12-01
#> 7     7            Rcpp 236297 2019-11-25 2019-12-01
#> 8     8           purrr 231133 2019-11-25 2019-12-01
#> 9     9         ggplot2 231039 2019-11-25 2019-12-01
#> 10   10        ellipsis 229053 2019-11-25 2019-12-01

# }