Top downloaded packages from the RStudio CRAN mirror

Top downloaded packages from the RStudio CRAN mirror

cran_top_downloads(when = c("last-day", "last-week", "last-month"),
  count = 10)

Arguments

when
last-day, last-week or last-month (see details).

count
Number of packages to list. Note that the DB server lists only at most 100 packages. This number might change in the future.

Value

A data frame with columns: rank, package, count, from, to.

Details

last-day is the last day for which data is available, last-week is from 6 days prior to that last day with data, last-month is from 29 days prior to that last day with data.

0 counts can be due to the non-availability of data on the RStudio server for that day.

See also

Examples

## Default is last day for which data is available.
cran_top_downloads()

#>    rank              package count       from         to
#> 1     1             devtools 30485 2019-04-27 2019-04-27
#> 2     2       microbenchmark 19588 2019-04-27 2019-04-27
#> 3     3                dplyr 19318 2019-04-27 2019-04-27
#> 4     4       assertive.base 19216 2019-04-27 2019-04-27
#> 5     5 assertive.reflection 19210 2019-04-27 2019-04-27
#> 6     6             fortunes 19204 2019-04-27 2019-04-27
#> 7     7          googledrive 19171 2019-04-27 2019-04-27
#> 8     8                rlang 18953 2019-04-27 2019-04-27
#> 9     9                 Rcpp 17224 2019-04-27 2019-04-27
#> 10   10              ggplot2 16977 2019-04-27 2019-04-27

## Last week (6 days prior to the last day for which data is available) 
## instead
cran_top_downloads(when = "last-week")

#>    rank    package  count       from         to
#> 1     1   devtools 234247 2019-04-21 2019-04-27
#> 2     2      rlang 210403 2019-04-21 2019-04-27
#> 3     3      dplyr 193425 2019-04-21 2019-04-27
#> 4     4       Rcpp 191614 2019-04-21 2019-04-27
#> 5     5    ggplot2 176468 2019-04-21 2019-04-27
#> 6     6     tibble 173132 2019-04-21 2019-04-27
#> 7     7 data.table 149969 2019-04-21 2019-04-27
#> 8     8       glue 141711 2019-04-21 2019-04-27
#> 9     9        cli 140817 2019-04-21 2019-04-27
#> 10   10     pillar 140412 2019-04-21 2019-04-27